Fix attendance calculation and input types

+++ kala.py
@@ -1,5 +1,5 @@
 total_classes = int(input("Enter the total number of classes: "))
-attended_classes = int(input("Enter the number of classes attended: "))
+attended_classes = float(input("Enter the number of classes attended: "))
 attendance_percentage = (attended_classes / total_classes) * 100
 print(f"Your attendance percentage is: {attendance_percentage:.2f}%")
 if("attendance percentage >75%"):
@@ -7,3 +7,4 @@ if("attendance percentage >75%"):
 else:
   print("status:not eligible")
     
+
